Fixed regression I added in 717ff81a9afd6e7ce795ea40a0572cc4e0e8d20f
When I renamed ivars to start with underscore,
I missed the fact that 'remoteBehavior' was bound to
from the nib. There was no 'remoteBehavior' accessor
so the bounding was made invalid.

Took this opportunity to use the first @property
in this codebase.  This is still compatible back to 10.5
and gcc 4.