Renamed two ivars to match their methods; changed nib to xib
Renamed _maxClickTimeDifference ivar to _maximumClickCountTimeDifference.
Renamed _simulateHoldEvents ivar to _simulateHoldEvent. (Note singular vs plural.)

This will help with eventually moving to auto-synthesize.

Used Xcode/IB 3.1.4 on OS 10.5 to convert the nib to xib; still deploys to OS X 10.5 and later.

Renamed binding in nib from simulateHoldEvents to simulateHoldEvent,
the binding was to the (previously misnamed) ivar, instead of the method,
probably broken when I added underscore prefixes to ivars.